Remove non-existing function call to fix segfault
In the implementation for the reset function, there is a call to a
function that doesn't exist in the current codebase. Calling reset(),
as a result, causes a segmentation fault every single time, making the
reset() function completely unusable. This commit removes that function
call and fixes the segmentation fault issue.